+++ lib/src/models/config_models/voice_message_configuration.dart
@@ -23,10 +23,16 @@
 import 'package:audio_waveforms/audio_waveforms.dart';
 import 'package:flutter/material.dart';
 
+import '../../values/typedefs.dart';
+
 /// A configuration model class for voice message bubble.
 class VoiceMessageConfiguration {
   const VoiceMessageConfiguration({
+    @Deprecated(
+        'Use `outgoingPlayerWaveStyle`/`inComingPlayerWaveStyle` instead.')
     this.playerWaveStyle,
+    this.outgoingPlayerWaveStyle,
+    this.inComingPlayerWaveStyle,
     this.padding = const EdgeInsets.symmetric(horizontal: 8),
     this.margin,
     this.decoration,
@@ -40,7 +46,16 @@ class VoiceMessageConfiguration {
   });
 
   /// Applies style to waveform.
+  ///
+  /// **Note:** if both [playerWaveStyle] and [outgoingPlayerWaveStyle]/[inComingPlayerWaveStyle] are provided,
+  /// then [outgoingPlayerWaveStyle]/[inComingPlayerWaveStyle] will take precedence.
   final PlayerWaveStyle? playerWaveStyle;
+
+  /// Applies style to outgoing waveform.
+  final PlayerWaveStyle? outgoingPlayerWaveStyle;
+
+  /// Applies style to incoming waveform.
+  final PlayerWaveStyle? inComingPlayerWaveStyle;
 
   /// Applies padding to message bubble.
   final EdgeInsets padding;
@@ -63,11 +78,11 @@ class VoiceMessageConfiguration {
   /// Curve for for grow animation for waveform. Default to Curve.easeIn.
   final Curve? animationCurve;
 
-  /// Icon for playing the audio.
-  final Icon? playIcon;
+  /// Callback to build custom icon for playing audio
+  final CustomVoiceActionIconCallback? playIcon;
 
-  /// Icon for pausing audio
-  final Icon? pauseIcon;
+  /// Callback to build custom icon for pausing audio
+  final CustomVoiceActionIconCallback? pauseIcon;
 
   /// Enable/disable seeking with gestures. Enabled by default.
   final bool enableSeekGesture;